session-replay: extract options to separate file, add inspect callback
